Arreglando controladores 20%

+++ app/Http/Controllers/AdministradoresController.php
@@ -8,7 +8,7 @@ use App\Http\Requests\AdministradoresRequest;
 
 class AdministradoresController extends Controller
 {
-
+    //Probada
     public function index()
     {
         return Administrador::all();
@@ -36,8 +36,12 @@ class AdministradoresController extends Controller
 
     public function show($id)
     {
-        $administrador = Administrador::find($id);
-        return $administrador;
+            $administrador = Administrador::find($id);
+            if ($administrador != NULL)
+                return $administrador;
+            else   
+                return "No existe un administrador con esa id";    
+        
     }
 
     public function edit(Administrador $administrador)
@@ -63,8 +67,16 @@ class AdministradoresController extends Controller
 
     public function destroy($id)
     {
+        
         $administrador = Administrador::find($id);
-        $administrador->delete();
-        return "El Administrador ha sido eliminado de la DB";
+
+        if($administrador != NULL){
+            $administrador->delete();
+            Administrador::destroy($id);
+            return "El Administrador ha sido eliminado de la DB";
+        }
+        else
+            return "No existe un administardor con el id ingresado";
+        
     }
 }
